+++ src/common/organisms/Navigation/index.tsx
@@ -7,7 +7,7 @@ import Toolbar from '@mui/material/Toolbar';
 import Divider from '@mui/material/Divider';
 import List from '@mui/material/List';
 import ListSubheader from '@mui/material/ListSubheader';
-import NavigationItem from 'src/common/molecules/NavigationItem';
+import NavigationItem from '../../molecules/NavigationItem';
 
 export default function Navigation() {
   return (
